Updated the build system for platforms; they are now seperated by architecture added an in-ROM filesystem driver added the ability to configure which drivers get which disk name added automounting a disk on startup preliminary x86 porting automated build system for userspace libraries automated generation of the ROM filesystem added more userspace demo programs
the ROM filesystem is generated from the local directory "root" in the base dir. everything put in here will
be included into the root filesystem image.
Currently there is an automated build system that compiles any C sources in the usr dir of the project base, and places them
into the usr dir of the rom filesystem root automatically before filesystem synthesis